+++ print_number.c
@@ -5,7 +5,7 @@
  * @n: Number
  * Return: count
  */
-int print_number(int n, const  char **format, int fc)
+int print_number(long int n, const  char **format, int fc)
 {
 	char s;
 	int c = 0;
